% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{create_badge}
\alias{create_badge}
\title{Create badge svg}
\usage{
create_badge(path, ...)
}
\arguments{
\item{path}{path to save svg to}

\item{...}{paramaters to pass to @seealso [Badge]}
}
\value{
svg text
}
\description{
Create badge svg
}
\examples{
\dontrun{
create_badge("anybadger.svg", label = "any", value = "badger", color = "brown")
}
}
