% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{Badge}
\alias{Badge}
\title{Add documentation}
\description{
Add documentation

Add documentation
}
\examples{

## ------------------------------------------------
## Method `Badge$new`
## ------------------------------------------------

b <- Badge$new()

## ------------------------------------------------
## Method `Badge$create_svg`
## ------------------------------------------------

\dontrun{
b <- Badge$new()
b$create_svg()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Badge$new()}}
\item \href{#method-create_svg}{\code{Badge$create_svg()}}
\item \href{#method-clone}{\code{Badge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Documentation here
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Badge$new(
  label = NULL,
  value = NULL,
  thresholds = NULL,
  font_name = NULL,
  font_size = NULL,
  label_text_color = NULL,
  value_text_color = NULL,
  color = NULL,
  svg_template_path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{todo}

\item{\code{value}}{todo}

\item{\code{thresholds}}{todo}

\item{\code{font_name}}{todo}

\item{\code{font_size}}{todo}

\item{\code{label_text_color}}{todo}

\item{\code{value_text_color}}{todo}

\item{\code{color}}{todo}

\item{\code{svg_template_path}}{todo}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
null
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{b <- Badge$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_svg"></a>}}
\subsection{Method \code{create_svg()}}{
Todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Badge$create_svg(path = "default_badge.svg")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{todo}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
todo
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
b <- Badge$new()
b$create_svg()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Badge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
